requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-multipart>=0.0.9
jinja2>=3.1.3
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.1.2
aiofiles>=23.2.1
cryptography>=42.0.0
requests>=2.31.0
